API/requirements.txt
Django==2.1.2
djangorestframework==3.9.0
djoser==1.3.1
django-filter==2.0.0